More GUI and 2-Phase Solver
Using an external 2-Phase solver written in C++.  Eventually would like to use it as a blue print to write a 2-Phase solver in Pascal however for now I am including the compiled binary of min2phase for Linux and just piping the solution to the solver.  Made many changes to the GUI, adjusted 3D rendering and 2d cube rendering.  Minor GUI improvements such as disabling buttons while a listg of commands are being executed on a cube.  Implemented keyboard control which needs to be refined.